backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
python-multipart==0.0.9

fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
anthropic          # Claude API
pypdf2             # PDF o'qish
python-docx        # .doc/.docx o'qish
sentence-transformers  # embedding (bepul, local)
numpy
pgvector         # PostgreSQL uchun vektor ma'lumotlar bazasi kengaytmasi
google-generativeai